+Initial working number detector ann

+++ src/ai/numberdetector/image/file/MNISTData.java
@@ -3,11 +3,11 @@ package ai.numberdetector.image.file;
 import java.util.List;
 
 public class MNISTData {
-    public List<int[]> images;
+    public List<double[]> images;
     public int rows;
     public int cols;
 
-    public MNISTData(List<int[]> images, int rows, int cols) {
+    public MNISTData(List<double[]> images, int rows, int cols) {
         this.images = images;
         this.rows = rows;
         this.cols = cols;
